requirements.txt
dj-database-url==0.5.0
Django==2.1.7
django-heroku==0.3.1
djangorestframework==3.9.2
gunicorn==19.9.0
Jinja2==2.10.1
MarkupSafe==1.1.1
msgpack==0.6.1
psycopg2==2.8.2
pytz==2019.1
Twisted==19.2.0
txaio==18.8.1
virtualenv==16.5.0
virtualenv-clone==0.5.3
whitenoise==4.1.2
zope.interface==4.6.0
